+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -1,5 +1,8 @@
 
 # Changelog
+## [0.9.018-c] - 2025-06-23
+### Fixed
+- Add a trade button typo
 
 ## [0.9.017-c] - 2025-06-22
 ### Added